SON-8
SON-8

+++ src/main/java/com/octo/sonar/demo/sampleproject/DemoCalculator.java
@@ -8,7 +8,7 @@ package com.octo.sonar.demo.sampleproject;
  *
  */
 public class DemoCalculator implements DemoCalculatorInterface {
-	private final static Integer abc = new Integer(1);
+	private final static Integer abc = new Integer(2);
 	
 	/**
 	 * This is the demo calculator class
